requirements.txt
Flask==2.3.2
google-cloud-datastore==2.8.2
python-jose
authlib
requests
